Split C codegen program generation tests
