+++ src/codegen/c/tests/expression_emission.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::ast::expressions::BinaryOp;
 
 #[test]
 fn emit_int_literal() {
