+++ src/codegen/c/tests/mod.rs
@@ -1,163 +1,7 @@
 use super::*;
-use crate::ast::expressions::BinaryOp;
 
+mod program_generation;
 mod type_mapping;
-
-fn make_simple_program() -> TypedProgram {
-    TypedProgram {
-        functions: vec![TypedFunction {
-            name: "add".into(),
-            params: vec![
-                TypedParam {
-                    name: "a".into(),
-                    ty: Type::I32,
-                    span: crate::error::Span::dummy(),
-                },
-                TypedParam {
-                    name: "b".into(),
-                    ty: Type::I32,
-                    span: crate::error::Span::dummy(),
-                },
-            ],
-            return_type: Type::I32,
-            body: TypedBlock {
-                statements: vec![],
-                expr: Some(Box::new(TypedExpression {
-                    kind: TypedExprKind::Return(Some(Box::new(TypedExpression {
-                        kind: TypedExprKind::BinaryOp {
-                            op: BinaryOp::Add,
-                            left: Box::new(TypedExpression {
-                                kind: TypedExprKind::Variable("a".into()),
-                                ty: Type::I32,
-                                span: crate::error::Span::dummy(),
-                            }),
-                            right: Box::new(TypedExpression {
-                                kind: TypedExprKind::Variable("b".into()),
-                                ty: Type::I32,
-                                span: crate::error::Span::dummy(),
-                            }),
-                        },
-                        ty: Type::I32,
-                        span: crate::error::Span::dummy(),
-                    }))),
-                    ty: Type::Never,
-                    span: crate::error::Span::dummy(),
-                })),
-                ty: Type::I32,
-                span: crate::error::Span::dummy(),
-            },
-            defers: vec![],
-            span: crate::error::Span::dummy(),
-        }],
-        types: vec![],
-        globals: vec![],
-        entry_point: None,
-    }
-}
-
-#[test]
-fn generates_function() {
-    let backend = CBackend;
-    let program = make_simple_program();
-    let output = backend.generate(&program).unwrap();
-    assert!(output.contains("int32_t add(int32_t a, int32_t b)"));
-    assert!(output.contains("return (a + b)"));
-}
-
-#[test]
-fn generates_struct() {
-    let backend = CBackend;
-    let program = TypedProgram {
-        functions: vec![],
-        types: vec![TypedTypeDef {
-            name: "Point".into(),
-            kind: TypeDefKind::Struct {
-                fields: vec![("x".into(), Type::F64), ("y".into(), Type::F64)],
-            },
-            methods: vec![],
-            span: crate::error::Span::dummy(),
-        }],
-        globals: vec![],
-        entry_point: None,
-    };
-    let output = backend.generate(&program).unwrap();
-    assert!(output.contains("typedef struct Point Point;"));
-    assert!(output.contains("double x;"));
-    assert!(output.contains("double y;"));
-}
-
-#[test]
-fn generates_enum() {
-    let backend = CBackend;
-    let program = TypedProgram {
-        functions: vec![],
-        types: vec![TypedTypeDef {
-            name: "Color".into(),
-            kind: TypeDefKind::Enum {
-                variants: vec![
-                    TypedVariant {
-                        name: "Red".into(),
-                        tag: 0,
-                        payload: None,
-                    },
-                    TypedVariant {
-                        name: "Green".into(),
-                        tag: 1,
-                        payload: None,
-                    },
-                    TypedVariant {
-                        name: "Blue".into(),
-                        tag: 2,
-                        payload: None,
-                    },
-                ],
-            },
-            methods: vec![],
-            span: crate::error::Span::dummy(),
-        }],
-        globals: vec![],
-        entry_point: None,
-    };
-    let output = backend.generate(&program).unwrap();
-    assert!(output.contains("Color_Red = 0"));
-    assert!(output.contains("Color_Green = 1"));
-    assert!(output.contains("Color_Blue = 2"));
-    assert!(output.contains("enum Color_Tag tag;"));
-}
-
-#[test]
-fn generates_entry_point() {
-    let backend = CBackend;
-    let program = TypedProgram {
-        functions: vec![TypedFunction {
-            name: "main".into(),
-            params: vec![],
-            return_type: Type::I32,
-            body: TypedBlock {
-                statements: vec![],
-                expr: Some(Box::new(TypedExpression {
-                    kind: TypedExprKind::Return(Some(Box::new(TypedExpression {
-                        kind: TypedExprKind::IntLiteral(0),
-                        ty: Type::I32,
-                        span: crate::error::Span::dummy(),
-                    }))),
-                    ty: Type::Never,
-                    span: crate::error::Span::dummy(),
-                })),
-                ty: Type::I32,
-                span: crate::error::Span::dummy(),
-            },
-            defers: vec![],
-            span: crate::error::Span::dummy(),
-        }],
-        types: vec![],
-        globals: vec![],
-        entry_point: Some("main".into()),
-    };
-    let output = backend.generate(&program).unwrap();
-    assert!(output.contains("int main(int argc, char** argv)"));
-    assert!(output.contains("return zen_main()"));
-}
 
 fn dummy() -> crate::error::Span {
     crate::error::Span::dummy()
@@ -173,77 +17,3 @@ fn texpr(kind: TypedExprKind, ty: Type) -> TypedExpression {
 
 mod expression_emission;
 mod helpers;
-
-#[test]
-fn generates_enum_with_payload() {
-    let backend = CBackend;
-    let program = TypedProgram {
-        functions: vec![],
-        types: vec![TypedTypeDef {
-            name: "Shape".into(),
-            kind: TypeDefKind::Enum {
-                variants: vec![
-                    TypedVariant {
-                        name: "Circle".into(),
-                        tag: 0,
-                        payload: Some(vec![("radius".into(), Type::I32)]),
-                    },
-                    TypedVariant {
-                        name: "Square".into(),
-                        tag: 1,
-                        payload: Some(vec![("side".into(), Type::I32)]),
-                    },
-                ],
-            },
-            methods: vec![],
-            span: dummy(),
-        }],
-        globals: vec![],
-        entry_point: None,
-    };
-    let output = backend.generate(&program).unwrap();
-    assert!(output.contains("Shape_Circle = 0"));
-    assert!(output.contains("Shape_Square = 1"));
-    assert!(output.contains("int32_t circle;"));
-    assert!(output.contains("int32_t square;"));
-}
-
-#[test]
-fn generates_function_with_defers() {
-    let backend = CBackend;
-    let program = TypedProgram {
-        functions: vec![TypedFunction {
-            name: "process".into(),
-            params: vec![],
-            return_type: Type::I32,
-            body: TypedBlock {
-                statements: vec![],
-                expr: Some(Box::new(texpr(
-                    TypedExprKind::Return(Some(Box::new(texpr(
-                        TypedExprKind::IntLiteral(42),
-                        Type::I32,
-                    )))),
-                    Type::Never,
-                ))),
-                ty: Type::I32,
-                span: dummy(),
-            },
-            defers: vec![texpr(
-                TypedExprKind::FunctionCall {
-                    function: "cleanup".into(),
-                    args: vec![],
-                },
-                Type::Void,
-            )],
-            span: dummy(),
-        }],
-        types: vec![],
-        globals: vec![],
-        entry_point: None,
-    };
-    let output = backend.generate(&program).unwrap();
-    // Defer should emit cleanup before return
-    assert!(output.contains("cleanup()"));
-    // Should save return value to temp, run defers, then return temp
-    assert!(output.contains("__tmp"));
-}
